Move MediaProgressIndicator to its own file
