+++ lib/src/media_player/multi_medium_player.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart' show ChangeNotifierProvider, Consumer;
 
 import 'media_player_error.dart' show MediaPlayerError;
+import 'media_progress_indicator.dart' show MediaProgressIndicator;
 import 'multi_medium_state.dart' show MultiMediumState, MultiMediumTrackState;
 
 /// A player for a [MultiMedium].
@@ -148,31 +149,4 @@ class MultiMediumTrackPlayer extends StatelessWidget {
           return MediaProgressIndicator(visualizable: state.isVisualizable);
         },
       );
-}
-
-/// A progress indicator that shows what kind of media is loading.
-///
-/// If [isVisualizable] is true, a [Icons.local_movies] will be shown, otherwise
-/// a [Icons.music_note] will be shown. A [CircularProgressIndicator] is always
-/// shown in the back.
-class MediaProgressIndicator extends StatelessWidget {
-  /// If true, a [Icons.local_movies] is shown, otherwise a [Icons.music_note].
-  final bool isVisualizable;
-
-  /// Constructs a [MediaProgressIndicator] setting if it's [visualizable].
-  const MediaProgressIndicator({@required bool visualizable})
-      : assert(visualizable != null),
-        isVisualizable = visualizable;
-
-  /// Builds the widget.
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      alignment: Alignment.center,
-      children: <Widget>[
-        Icon(isVisualizable ? Icons.local_movies : Icons.music_note),
-        CircularProgressIndicator(),
-      ],
-    );
-  }
 }
